Organized pipeline tests in their own dir
