+++ .pre-commit-config.yaml
@@ -35,6 +35,7 @@ repos:
         additional_dependencies:
           - types-PyYAML==5.4.10
           - types-setuptools==57.4.0
+        exclude: tests
   - repo: https://github.com/igorshubovych/markdownlint-cli
     rev: v0.29.0
     hooks:
